Move function get_fake_args to test utils
